DocumentTile used in both PdfTile and TextTile
