+++ libs/search-widget/src/tiles/base-tile/DocumentTile.scss
