--- libs/search-widget/src/tiles/text-tile/TextTile.scss
@@ -1,6 +1,0 @@
-@import '../tile';
-@import '../text-tile';
-
-.resource-viewer-container {
-  overflow: auto;
-}
